+++ src/main/java/SecondRoll/demo/services/GameAdsService.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
 @Service
-public class GameAdService {
+public class GameAdsService {
     @Autowired
     GameAdsRepository gameAdsRepository;
     // Create a gameAd
